edit 24

+++ data/natural_loader.py
@@ -33,6 +33,6 @@ class NaturalDatasetLoader(data.Dataset):
         print("Shape: ", img.shape)
         img = np.array(img, dtype=np.uint8)
         img = np.transpose(img, (2,0,1))
-
+        img = img.astype('float32')
         img_name = img_path.split('/')[-1]
         return img, img_path, img_name
